requirements.txt
screeninfo
panda
panda3d
pygame-ce
pygame_gui